switched to postgres database for development and test environment
switched to postgres database for development and test environment. This
is to make sure that we have the same database backend for both
development and production. Also the TURN server code in C can be
written to make use of postgres so that the same can be used in both
production and development.