Added displaying of allocation table in a tabular way using dataTables
In this commit, following are the changes made

1. Integrated the dataTables gem for rendering of a user's allocations
in a more tabular way. This gem basically works on the client side and
provides additional benefit of searching within the table, pagination
and sorting based on the columns. As mentioned, this takes all the
allocation table info from the server at one time and can be too much in
case the number of allocations is too high. In such a case, it is better
to make use of ajax and serve the allocations in parts using json. This
is supported in dataTables gem but the support for the same has not been
added in this commit. It is definitely a TODO.
2. Spruced up the table look and feel using the bootstrap classes and
also modified the css slightly.