Fix to avoid the looping when password reset feature is run.
When user opts to reset the password, a confirmation mail is sent to his
mail id with the url to reset the password. On this url when user
provides the new password and submits the form, there was a recursive
rendering of the edit password url. In order to fix this, just reordered
the url, not sure if this is the right way to fix it.